move encoding out of address package
